feat: Question 컨트롤러 & Dto 수정

+++ server/src/main/java/com/lucky7/preproject/question/dto/responseDto/AllQuestionsResponseDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AllQuestionsResponseDto {
-    private int questionId;
+    private long questionId;
     private String questionTitle;
     private String questionContent;
     private String questionUser;
